Add the etcd-arm64.yaml for supporting arm64
Add the etcd-arm64.yaml file on v3.5 & master version.
Firstly, the base template of etcd.yaml which supports
the different architecture by the value of "arch" was
created in _includes folder.
Secondly, the etcd-arm64.yaml was created in master &
v3.5 folders.
On the other hand, legacy etcd.yaml was modified by
supporting it.

Signed-off-by: Jingzhao <Jingzhao.Ni@arm.com>